AD lab 2
After a week of attempting to create the lab environment on VMWare finally was able to make it work on Virtual Box. Find the instructions in the document! 
The document was created with the help of Gemini after almost 2 hours of prompting and compilation ( GPT makes our life easier and frustrating at the same time)